Synchronize pod network operations
The runtime calling CRI-O may not synchronize pod network operations,
but to ensure that they don't overlap, ocicni must do so.  Operations
are synchronized on a per-pod basis, such that all operations for a
given pod are synchronized, but individual pods' operations can
run in parallel.

This is essentially a port of the Kubernetes PluginManager code from
https://github.com/kubernetes/kubernetes/pull/37036.

Signed-off-by: Dan Williams <dcbw@redhat.com>