ocicni: always monitor net dir
Currently, the net dir (usually `/etc/cni/net.d`) is only monitored for
new network configuration files during runtime iff it does not contain
any files during CNI initialization (`InitCNI`):

* If a valid configuration file is found during initialization, only
  exactly this plugin will be loaded and used throughout.
* If the directory contains only invalid configuration files or in the
  event on an error, a noop plugin is returned.

It's not unusual though for the net dir to contain default network
configuration or or fallback at first (for example when the cri-o
systemd unit is started) and to be updated later (for example by a
network plugin like flannel). At present, the update of the directory
would not be detected since the monitor routine is not started in any
case.

Instead, always start the monitor routine and never return a noop
plugin. This way new default configuration will be picked up immediately
or, if containing errors, lead to an error message in the log.

Signed-off-by: Michael Schubert <michael@kinvolk.io>