Update logging to be better traceable and more uniform
This commit updates the logging by adding a new `log()` function, which
adds the `origin=OCICNI` fields to all logs. This enables us to hint to
this library if a log occurs in API consumers. Beside this, most logs
start now with a lowercase letter and the overall verbosity has been
decreased (info → debug).

Signed-off-by: Sascha Grunert <sgrunert@suse.com>