ocicni: find networks to delete from cache
Instead of requiring the runtime to save network attachment state
for correct network teardown, allow the runtime to pass an empty
list of networks. Passing an empty list reads the attachments
from cache and tears them down using that information.

I anticipate the read-all-networks-from-cache functionality will
make it upstream to CNI soon, but for now keep it in ocicni.

Signed-off-by: Dan Williams <dcbw@redhat.com>