ocicni: return CNI Result and added interface name
ocicni generates the container interface name for each network
attachment, so the runtime doesn't always know what the interface
name is. And because CNI specification versions < 0.3.0 didn't
include the pod interface name in the Result, it's not always
possible for the runtime to determine the interface name from
the Result unless ocicni helps out.

Signed-off-by: Dan Williams <dcbw@redhat.com>