ocicni: use cached config for DEL/CHECK when possible
Changes to on-disk CNI config files may change CNI network
configuration underneath containers, causing DEL/CHECK operations
to be performed with different configuration and RuntimeConf than
the container was ADD-ed with.

To prevent this situation (and enable runtime changes to the
default network) load container configuration and RuntimeConf
from the CNI cache when possible.

Signed-off-by: Dan Williams <dcbw@redhat.com>