Remove error logging for deleteFromNetwork call
While working on podman#8571 I noticed that calling `deleteFromNetwork`
logs the same error twice. For example it errors if the iptables rules do
not exists and this creates a very verbose error. Since the error is returned
I do not see any point in logging it. The caller should decide if the
error should be logged or not.

It is super confusing for users to get a error logged when they have no
control to change anything. In case of podman#8571 the user would only want
to call this if the iptables rules are flushed, getting a error in this case
is super confusing.

Signed-off-by: Paul Holzinger <paul.holzinger@web.de>