Add InitCNIWithCache to set the cni cache directory
Add a new function InitCNIWithCache which works like
InitCNI except that it accepts an extra parameter to
set the cache directory.

This change is needed for rootless users which are
trying to use cni in a user+net namespace because the
default directory is not writeable by rootless users.

Second, setting the cni cache directory via the runtime
config is deprecated. It has to be set with
NewCNIConfigWithCacheDir for the cni config.
see: containernetworking/cni#682

Signed-off-by: Paul Holzinger <paul.holzinger@web.de>