Reload networks if none is found
In the podman CI tests are consistently flaking with CNI network not
found ... I think the problem is that ocicni updates based on fsnotify
and this can be slower than the podman calls.
As a solution the networks should be synced if no match is found. If
it is still not found afterwards return the network not found error.

Signed-off-by: Paul Holzinger <paul.holzinger@web.de>