ocicni: watch bindirs too
Config validation involves finding and execing the config's plugin
to determine the CNI spec versions the plugin binary supports.

If the config is written to disk before or concurrently with the
plugin binary then the validation may race with the binary and
fail to either find it or exec it.

Watch bindirs for changes and resync network config when plugin
binaries change.

Signed-off-by: Dan Williams <dcbw@redhat.com>