pkg/ocicni: Use 'ifconfig -j' to access jail network state
The use of 'jexec' for this requires a compatible ifconfig binary inside
the jail which owns the network state and using 'ifconfig -j' lets us
merge the jail which owns the pod network with the infra container.

This also fixes some parsing bugs in getContainerDetails which were not
noticed before since most of the time we get the information from cni's
CheckNetworkList.

Signed-off-by: Doug Rabson <dfr@rabson.org>