add GC support
This introduces CNI GC support, a mode in which the runtime provides a
list of expected attachments, and then network plugins can clean up stale
resources such as IPAM attachments.

Because CNI GC is not allowed to take place in parallel with any other
pod operations, a new top-level lock was added. I did not use the plugin
lock, as GC should not block non-pod calls such as STATUS.

Signed-off-by: Casey Callendrello <c1@caseyc.net>